% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{action_codes}
\alias{action_codes}
\title{Get Action Codes Mapping}
\usage{
action_codes()
}
\value{
A named list where each integer code is mapped to its action description.
}
\description{
This utility function returns a list that maps action codes to their corresponding descriptions in the \code{interview__actions} file from Survey Solutions.
}
