% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paradata.R
\name{analyze_para}
\alias{analyze_para}
\title{Analyze Paradata}
\usage{
analyze_para(dt = NULL, answ.changes = TRUE, duration = TRUE)
}
\arguments{
\item{dt}{A data.table object containing the paradata to be analyzed.}

\item{answ.changes}{Logical. If \code{TRUE}, identifies changes made to answers after the first entry. Default is \code{TRUE}.}

\item{duration}{Logical. If \code{TRUE}, computes the duration of events in seconds. Default is \code{TRUE}.}
}
\value{
A data.table with analyzed paradata, including identified answer changes and computed durations.
}
\description{
This function analyzes paradata from a data.table, identifying answer changes and computing event durations.
}
\details{
The function performs the following:
\itemize{
\item Checks if required columns are present.
\item Identifies changes to answers (if \code{answ.changes} is \code{TRUE}).
\item Computes durations of events (if \code{duration} is \code{TRUE}).
}
}
\examples{
\dontrun{
dt <- read_para("path/to/paradata")
analyzed_dt <- analyze_para(dt)
}
}
