% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interviewer_tablet_report.R
\name{get_int_report}
\alias{get_int_report}
\title{Get Interviewer Report from Survey Solutions Server}
\usage{
get_int_report(
  workspace = "primary",
  server = NULL,
  hquser = NULL,
  password = NULL,
  archived = FALSE,
  type = "tab"
)
}
\arguments{
\item{workspace}{A character string specifying the workspace name. Default is "primary".}

\item{server}{A character string specifying the server URL. This parameter is required.}

\item{hquser}{A character string specifying the Headquarters user name. This parameter is required.}

\item{password}{A character string specifying the password for the Headquarters user. This parameter is required.}

\item{archived}{A logical value indicating whether to include archived interviewers in the report. Default is \code{FALSE}.}

\item{type}{A character string specifying the format of the report. Must be one of "tab" or "csv". Default is "tab".}
}
\value{
A \code{data.table} containing the interviewer report.
}
\description{
This function retrieves a report of all interviewers from the specified
Survey Solutions server. The report can be returned in either tab-separated
values or CSV format.
}
\details{
This function constructs an API request to retrieve a list of all interviewers
from the specified Survey Solutions server and workspace. The report can be
returned in either tab-separated or CSV format, depending on the \code{type} parameter.

The function checks for the required arguments (\code{server}, \code{hquser}, and \code{password}),
and will stop with an error message if any of these are missing.
}
\examples{
\dontrun{
report <- get_int_report(
  workspace = "primary",
  server = "https://your-server-url/",
  hquser = "your-username",
  password = "your-password",
  archived = FALSE,
  type = "csv"
)
print(report)
}

}
