% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{label_interview_status}
\alias{label_interview_status}
\title{Convert and Label Interview Status Column}
\usage{
label_interview_status(dt)
}
\arguments{
\item{dt}{A \code{data.table} containing the data with an \code{interview__status} column.}
}
\value{
The input \code{data.table} with the \code{interview__status} column converted to a factor with descriptive labels.
}
\description{
This function converts the \code{interview__status} column in a \code{data.table} to a factor with descriptive labels.
It also checks that the column contains only valid status codes.
}
