% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{label_role}
\alias{label_role}
\title{Convert Role Column to Factor and Validate}
\usage{
label_role(dt)
}
\arguments{
\item{dt}{A \code{data.table} containing the data with a \code{role} column.}
}
\value{
The input \code{data.table} with the \code{role} column converted to a factor with descriptive labels.
}
\description{
This function converts the \code{role} column in a \code{data.table} to a factor with descriptive labels.
It also checks that the column contains only valid role codes. Usefol for Interview Comments
}
