% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{suso.sys.colClasses}
\alias{suso.sys.colClasses}
\title{Column Classes for SurSol System Generated Files}
\usage{
suso.sys.colClasses()
}
\value{
A list with three elements: \code{integer}, \code{Date} and \code{character}, each containing column names.
}
\description{
Returns a list of columns to be parsed as integer, date or character in \code{data.table::fread}.
}
\examples{
col_classes <- suso.sys.colClasses()
}
